Update deprecated php type casts
In PHP 8.5, non-canonical scalar type casts (boolean|double|integer|binary) are deprecated.

See https://php.watch/versions/8.5/boolean-double-integer-binary-casts-deprecated

+++ CRM/I3val/Handler/DetailUpdate.php
@@ -210,7 +210,7 @@ abstract class CRM_I3val_Handler_DetailUpdate extends CRM_I3val_ActivityHandler 
     }
 
     if ($similarity_count) {
-      return ((double) $similarity_sum / (double) $similarity_count);
+      return ((float) $similarity_sum / (float) $similarity_count);
     } else {
       return 0;
     }
